delete unneccessary scripts and game objects
