+++ Assets/Scripts/NPCDialogue.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: fb74f9fc4cae34d98adceadb770f4db2
+guid: 424d438b0432e4eabb397afcae709a81
