--- Assets/Scripts/NPCManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 424d438b0432e4eabb397afcae709a81
